requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==1.0.1
Sphinx==1.2.1
django-contrib-comments==1.5
